backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.1
openai==0.28
